Move cnl::set_digits into a separate header file
